backend/requirements.txt
annotated-types==0.7.0
astroid==4.0.4
bandit==1.9.4
blinker==1.9.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
dill==0.4.1
flake8==7.3.0
Flask==3.1.3
flask-cors==6.0.2
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
httplib2==0.31.2
idna==3.13
isort==8.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
platformdirs==4.9.6
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pyflakes==3.4.0
Pygments==2.20.0
pylint==4.0.5
pyparsing==3.3.2
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
rich==15.0.0
stevedore==5.7.0
tomlkit==0.14.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.8
